Research_Assistant/requirement.txt
langchain
langchain-core
langchain-community
langchain-huggingface
langgraph
pydantic
transformers
duckduckgo-search
arxiv
torch
python-dotenv